Add stm32f4 build
Add stm32f401 config build
bux fix for powerloss
bux fix for eeprom wirte time

+++ Marlin/src/feature/spindle_laser_types.h
@@ -42,7 +42,9 @@ typedef IF<(SPEED_POWER_MAX > 255), uint16_t, uint8_t>::type cutter_cpower_t;
   #define CUTTER_MENU_POWER_TYPE   uint16_5
   #define cutter_power2str         ui16tostr5rj
 #else
-  typedef uint8_t cutter_power_t;
+  // 107011 -20211008
+  //typedef uint8_t cutter_power_t;
+  typedef uint16_t cutter_power_t;
   #if CUTTER_UNIT_IS(PERCENT)
     #define CUTTER_MENU_POWER_TYPE percent_3
     #define cutter_power2str       pcttostrpctrj
